Converted TopHeader into stateless component, fetching userInfo from Panel component
