+++ src/main/frontend/src/components/header/TopHeader.js
@@ -1,10 +1,5 @@
-import React, { Component, Fragment } from 'react';
-import { Container, Row, Col } from 'reactstrap';
-import { connect } from 'react-redux';
-import { bindActionCreators } from 'redux';
+import React from 'react';
 import UserInfo from '../UserInfo';
-import { getUser } from '../../store/selectors';
-import { requestUser } from '../../store/actions/userActions';
 import ExecomLogo from '../../img/execom_logo.png';
 import styled from 'styled-components';
 
@@ -25,39 +20,13 @@ const LogoImage = styled.img`
 const NavSpan = styled.span`
   width: 100px;
 `;
-class TopHeader extends Component {
-  componentDidMount() {
-    this.props.requestUser(sessionStorage.getItem('userEmail'));
-  }
-  render() {
-    return (
-      <Fragment>
-        <NavHeader>
-          <NavSpan />
-          <LogoImage src={ExecomLogo} />
-          <UserInfo
-            fullName={this.props.user.fullName}
-            userEmail={this.props.user.email}
-          />
-        </NavHeader>
-      </Fragment>
-    );
-  }
-}
 
-const mapStateToProps = state => ({
-  user: getUser(state)
-});
+const TopHeader = props => (
+  <NavHeader>
+    <NavSpan />
+    <LogoImage src={ExecomLogo} />
+    <UserInfo fullName={props.user.fullName} userEmail={props.user.email} />
+  </NavHeader>
+);
 
-const mapDispatchToProps = dispatch =>
-  bindActionCreators(
-    {
-      requestUser
-    },
-    dispatch
-  );
-
-export default connect(
-  mapStateToProps,
-  mapDispatchToProps
-)(TopHeader);
+export default TopHeader;
